server/requirements.txt
flask
selenium
python-dotenv